# perf(transport): don't rotate_left then truncate stream send buffer