@martinthomson 2025-06-01
We shouldn't need to rotate, I agree.  `truncate_front()` will be ideal when it stabilizes.  Rust hasn't been very fast about that lately though.