@mxinden 2025-06-02
Looking at the `std::collection::VecDeque` implementation, I underestimated the power of a ring buffer.

Given the circular nature of a ring buffer our usage of `rotate_left` does not require a memory copy in most cases. Instead, `VecDeque::rotate_left` simply moves its front and back pointer.

That explains why `VecDeque::rotate_left` beats `VecDeque::drain`.

https://doc.rust-lang.org/src/alloc/collections/vec_deque/mod.rs.html#275-407

I will close here. Once `truncate_front` is available in Rust stable, we can reconsider a change. Though given that `VecDeque::rotate_left` is so efficient, a change to `truncate_front` will probably only make things easier to read, not actually faster.