@martinthomson 2025-06-02
@mxinden Did you open an issue to track `truncate_front` ?